New shapefile, var definitions
The shapefile is updated. Now, when you select a variable, its definition will appear below the plot area. And since the new shapefile contains much more details than the previous one. The multivariable barplot exhibition will be very hard to understand and slow to render. So I temperorily restrict the maximum number of CT with barplot to be 100. After I implement the zoom in, zoom out functionalities, this will be imporved.